+++ ft_memcpy.c
@@ -6,21 +6,20 @@
 /*   By: jlacerda <jlacerda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 19:56:49 by jlacerda          #+#    #+#             */
-/*   Updated: 2024/10/15 20:46:03 by jlacerda         ###   ########.fr       */
+/*   Updated: 2024/10/16 20:24:45 by jlacerda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include	"libft.h"
 
-void *ft_memcpy(void *dest, const void *src, unsigned long n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned long	i;
-	char			*ptr_dest;
-	char			*ptr_src;
+	size_t	i;
+	char	*ptr_dest;
+	char	*ptr_src;
 
 	if (!dest && !src)
 		return (NULL);
-
 	i = 0;
 	ptr_dest = (char *)dest;
 	ptr_src = (char *)src;
