+++ ft_memset.c
@@ -6,14 +6,16 @@
 /*   By: jlacerda <jlacerda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 19:56:49 by jlacerda          #+#    #+#             */
-/*   Updated: 2024/10/08 20:28:42 by jlacerda         ###   ########.fr       */
+/*   Updated: 2024/10/16 20:25:10 by jlacerda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned long n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned long	i;
-	char			*ptr;
+	size_t	i;
+	char	*ptr;
 
 	i = 0;
 	ptr = (char *)s;
